V.0.0.2.001 : Fix.
- Removed all byte convertion.
- Some value have been replaced by enum.
- MapData no longuer care of tile specifique properties, it add or remove tile only.
- MapData no longuer have his own appState, HexGridManager got the main reference.
- All listener are encapsulated, o longuer class implementation. (W.I.P)
- MouseEvent now use only one method for input, event contain the mouse input used by the user.

+++ HexGrid_API/src/org/hexgridapi/core/HexSetting.java
@@ -33,10 +33,10 @@ public final class HexSetting {
      * to handle negative value, mainly help with the meshManager and how
      * geometry is handle, generated.
      */
-    public final static byte GROUND_HEIGHT = 5;
+    public final static int GROUND_HEIGHT = 5;
     /**
      * Used to know how many chunk to keep in memory before start deleting.
      * Unused for the time being.
      */
-    public final static byte CHUNK_DATA_LIMIT = 4;
+    public final static int CHUNK_DATA_LIMIT = 4;
 }
